Bugfixes & Features added
Bug fixes:

Fixed a bug which caused to close the editor when the user tried to
paste text into it.

Fixed a bug which caused that the copy process didn’t worked as expected

Fixed a bug which within the loading process of a saved editor
instance. The wrong style settings were applied.

Added functionality:

Improved the save feature. Save only asks the first time for the path.
Save as asks every time for the path.

The name of the current file will be shown as a part of the window name.